orchestrator/requirements.txt
pandas
numpy
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
Faker
openai
tenacity
python-dotenv